Potion & Coin

+++ core/src/com/mygdx/game/src/World/ThreadWorld.java
@@ -13,10 +13,10 @@ public class ThreadWorld extends Thread {
 		while (true) {
 			long attuale = System.currentTimeMillis();
 			float dt = (float)(attuale - start);
-			//System.out.println(dt);
-			game.world.update(dt/1000);
+			
+			game.world.update((float) dt/1000);
 			start = attuale;
-			//System.out.println(this.getId());
+			
 		}
 	}
 }
